Move CLI stuff out of main package for increased testability. Refs #2313
