+++ cmd/influx/cli/cli_test.go
@@ -1,4 +1,4 @@
-package main_test
+package cli_test
 
 import (
 	"encoding/json"
@@ -8,19 +8,18 @@ import (
 	"testing"
 
 	"github.com/influxdb/influxdb/client"
-	main "github.com/influxdb/influxdb/cmd/influx"
+	"github.com/influxdb/influxdb/cmd/influx/cli"
 )
 
 func TestParseCommand_CommandsExist(t *testing.T) {
 	t.Parallel()
-	c := main.CommandLine{}
+	c := cli.CommandLine{}
 	tests := []struct {
 		cmd string
 	}{
 		{cmd: "gopher"},
 		{cmd: "connect"},
 		{cmd: "help"},
-		{cmd: "history"},
 		{cmd: "pretty"},
 		{cmd: "use"},
 		{cmd: ""}, // test that a blank command just returns
@@ -32,45 +31,9 @@ func TestParseCommand_CommandsExist(t *testing.T) {
 	}
 }
 
-func TestParseCommand_CommandsSamePrefix(t *testing.T) {
-	t.Parallel()
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		var data client.Response
-		w.WriteHeader(http.StatusNoContent)
-		_ = json.NewEncoder(w).Encode(data)
-	}))
-	defer ts.Close()
-
-	u, _ := url.Parse(ts.URL)
-	config := client.Config{URL: *u}
-	c, err := client.NewClient(config)
-	if err != nil {
-		t.Fatalf("unexpected error.  expected %v, actual %v", nil, err)
-	}
-	m := main.CommandLine{Client: c}
-
-	tests := []struct {
-		cmd string
-	}{
-		{cmd: "use db"},
-		{cmd: "user nodb"},
-		{cmd: "puse nodb"},
-		{cmd: ""}, // test that a blank command just returns
-	}
-	for _, test := range tests {
-		if !m.ParseCommand(test.cmd) {
-			t.Fatalf(`Command failed for %q.`, test.cmd)
-		}
-	}
-
-	if m.Database != "db" {
-		t.Fatalf(`Command "use" changed database to %q. Expected db`, m.Database)
-	}
-}
-
 func TestParseCommand_TogglePretty(t *testing.T) {
 	t.Parallel()
-	c := main.CommandLine{}
+	c := cli.CommandLine{}
 	if c.Pretty {
 		t.Fatalf(`Pretty should be false.`)
 	}
@@ -86,7 +49,7 @@ func TestParseCommand_TogglePretty(t *testing.T) {
 
 func TestParseCommand_Exit(t *testing.T) {
 	t.Parallel()
-	c := main.CommandLine{}
+	c := cli.CommandLine{}
 	tests := []struct {
 		cmd string
 	}{
@@ -105,7 +68,7 @@ func TestParseCommand_Exit(t *testing.T) {
 
 func TestParseCommand_Use(t *testing.T) {
 	t.Parallel()
-	c := main.CommandLine{}
+	c := cli.CommandLine{}
 	tests := []struct {
 		cmd string
 	}{
@@ -130,7 +93,7 @@ func TestParseCommand_Use(t *testing.T) {
 
 func TestParseCommand_Consistency(t *testing.T) {
 	t.Parallel()
-	c := main.CommandLine{}
+	c := cli.CommandLine{}
 	tests := []struct {
 		cmd string
 	}{
@@ -168,7 +131,7 @@ func TestParseCommand_Insert(t *testing.T) {
 	if err != nil {
 		t.Fatalf("unexpected error.  expected %v, actual %v", nil, err)
 	}
-	m := main.CommandLine{Client: c}
+	m := cli.CommandLine{Client: c}
 
 	tests := []struct {
 		cmd string
@@ -205,7 +168,7 @@ func TestParseCommand_InsertInto(t *testing.T) {
 	if err != nil {
 		t.Fatalf("unexpected error.  expected %v, actual %v", nil, err)
 	}
-	m := main.CommandLine{Client: c}
+	m := cli.CommandLine{Client: c}
 
 	tests := []struct {
 		cmd, db, rp string
@@ -257,7 +220,7 @@ func TestParseCommand_InsertInto(t *testing.T) {
 
 func TestParseCommand_History(t *testing.T) {
 	t.Parallel()
-	c := main.CommandLine{}
+	c := cli.CommandLine{}
 	tests := []struct {
 		cmd string
 	}{
